Expose monitorDataOutput hook to plugin library
- Allow plugins to process data outputs
