+++ src/frontend/package.json
@@ -1,7 +1,7 @@
 {
     "name": "@inventreedb/ui",
     "description": "UI components for the InvenTree project",
-    "version": "0.7.0",
+    "version": "0.8.0",
     "private": false,
     "type": "module",
     "license": "MIT",
